Registry integration tests: Add test tags.
