+++ tests/integration/test_fim/test_registry/test_tags/data/wazuh_registry_tag_conf.yaml
@@ -1,0 +1,21 @@
+---
+# Registry tags configuration
+- tags:
+  - ossec_conf
+  apply_to_modules:
+  - test_registry_tags
+  sections:
+  - section: syscheck
+    elements:
+    - disabled:
+        value: 'no'
+    - windows_registry:
+        value: WINDOWS_REGISTRY_1
+        attributes:
+          - arch: both
+          - tags: FIM_TAGS
+    - windows_registry:
+        value: WINDOWS_REGISTRY_2
+        attributes:
+          - arch: 64bit
+          - tags: FIM_TAGS
